+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Fix filtered drivers when update [#111](https://github.com/DevAlexandreCR/admin-driver/pull/111)
+
 # Release Notes for 1.2.x
 
 ## [1.2.9(2023-09-19)](https://github.com/DevAlexandreCR/admin-driver/compare/1.2.9...1.2.8)