+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Now admin can set 2 or more whatsapp clients. [#112](https://github.com/DevAlexandreCR/admin-driver/pull/112)
 
+### Changed
+- Add responsive tabs to dashboard. [#107](https://github.com/DevAlexandreCR/admin-driver/pull/107)
+
 # Release Notes for 1.2.x
 
 ## [1.2.9(2023-09-19)](https://github.com/DevAlexandreCR/admin-driver/compare/1.2.9...1.2.8)