+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fix filtered drivers when update [#111](https://github.com/DevAlexandreCR/admin-driver/pull/111)
 
+### Changed
+- Add responsive tabs to dashboard. [#107](https://github.com/DevAlexandreCR/admin-driver/pull/107)
+
 # Release Notes for 1.2.x
 
 ## [1.2.9(2023-09-19)](https://github.com/DevAlexandreCR/admin-driver/compare/1.2.9...1.2.8)